Emit update now event from OutputEventRenderer
The synchronization of the renderer pipeline is already done here.
